Rename the Custom3 test artifact to file and Adding doc comments for Capability
