+++ tests/pom.xml
@@ -110,7 +110,7 @@
             </dependency>
             <dependency>
                 <groupId>org.wso2.carbon</groupId>
-                <artifactId>org.wso2.carbon.sample.transport.custom3</artifactId>
+                <artifactId>org.wso2.carbon.sample.transport.file</artifactId>
                 <version>${carbon.kernel.version}</version>
             </dependency>
             <dependency>
