+++ tests/test-artifacts/startup-coordination/org.wso2.carbon.sample.transport.file/src/main/java/org/wso2/carbon/sample/transport/file/CustomTransportServiceCapabilityProvider.java
@@ -1,4 +1,19 @@
-package org.wso2.carbon.sample.transport.custom3;
+/*
+ *  Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+package org.wso2.carbon.sample.transport.file;
 
 import org.osgi.framework.BundleContext;
 import org.osgi.service.component.annotations.Activate;
@@ -24,7 +39,7 @@ public class CustomTransportServiceCapabilityProvider implements CapabilityProvi
     protected void start(BundleContext bundleContext) {
         Thread thread = new Thread(() -> {
             for (int i = 0; i < customTransportServiceCount; i++) {
-                bundleContext.registerService(Transport.class, new Custom3Transport(), null);
+                bundleContext.registerService(Transport.class, new FileTransport(), null);
 
                 try {
                     Thread.sleep(1000 * 2);
