+++ tests/test-artifacts/startup-coordination/pom.xml
@@ -33,7 +33,7 @@
         <module>org.wso2.carbon.sample.transport.jetty</module>
         <module>org.wso2.carbon.sample.transport.custom</module>
         <module>org.wso2.carbon.sample.transport.custom2</module>
-        <module>org.wso2.carbon.sample.transport.custom3</module>
+        <module>org.wso2.carbon.sample.transport.file</module>
         <module>org.wso2.carbon.sample.transport.mgt</module>
         <module>org.wso2.carbon.sample.runtime.mgt</module>
         <module>org.wso2.carbon.sample.runtime.mss</module>
